+++ tests/Service/TypeDossier/TypeDossierDeletionServiceTest.php
@@ -35,9 +35,9 @@ class TypeDossierDeletionServiceTest extends PastellTestCase
             $journal_raw_data
         );
 
-        $log_message = $this->getLogRecords()[1]['message'];
+        $log_message = $this->getLogRecords()[0]['message'];
         $this->assertRegExp(
-            "#Ajout au journal \(id_j=2\): 12 - 0 - 1 - 0 - Supprimé - Suppression du type de dossier id_t=1#",
+            "#Ajout au journal \(id_j=1\): 12 - 0 - 1 - 0 - Supprimé - Suppression du type de dossier id_t=1#",
             $log_message
         );
     }